requirements.txt
imap-tools
supabase
google-generativeai
python-dotenv
beautifulsoup4
line-bot-sdk